+++ include/quill/Logger.h
@@ -96,21 +96,11 @@ public:
                                 .count())
       : user_clock->now();
 
-    #if defined(__GNUC__) && defined(__linux__)
-    // On GCC (or compatible compilers) running on Linux, use __thread for thread-local storage.
-    // This optimizes performance by caching the ThreadContext pointer to avoid repeatedly
-    // calling get_local_thread_context(). If the thread_context is not already set, we initialize it.
     if (QUILL_UNLIKELY(thread_context == nullptr))
     {
+      // This caches the ThreadContext pointer to avoid repeatedly calling get_local_thread_context()
       thread_context = detail::get_local_thread_context<frontend_options_t>();
     }
-    #else
-    // On other compilers or platforms, use thread_local to store the ThreadContext pointer.
-    // This ensures portability across different compilers and operating systems.
-    // We fetch the ThreadContext pointer each time by calling get_local_thread_context().
-    detail::ThreadContext* const thread_context =
-      quill::detail::get_local_thread_context<frontend_options_t>();
-    #endif
 
     // Need to know how much size we need from the queue
     size_t total_size = sizeof(current_timestamp) + (sizeof(uintptr_t) * 3) +
@@ -299,8 +289,8 @@ public:
     std::atomic<bool>* backend_thread_flushed_ptr = &backend_thread_flushed;
 
     // We do not want to drop the message if a dropping queue is used
-    while (!this->log_statement<false>(
-      LogLevel::None, &macro_metadata, reinterpret_cast<uintptr_t>(backend_thread_flushed_ptr)))
+    while (!this->log_statement<false>(LogLevel::None, &macro_metadata,
+                                       reinterpret_cast<uintptr_t>(backend_thread_flushed_ptr)))
     {
       if (sleep_duration_ns > 0)
       {